pyproject.toml
[tool.poetry]
name = "events"
version = "0.1.0"
description = ""
authors = ["number92 <rav.92@list.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
Pillow = "^9.5"
djoser = "^2.2.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
djangorestframework-simplejwt = "^5.3.1"
# channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.12.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''